+++ Sheeeeeeeeet/Items/Items/ActionSheetCollectionItem.swift
@@ -72,10 +72,9 @@ open class ActionSheetCollectionItem<T>: ActionSheetItem, UICollectionViewDataSo
     }
     
     open func extendSelectionAction(toReload actionSheet: ActionSheet) {
-        weak var sheet = actionSheet
-        extendSelectionAction(with: { _, _ in
-//            sheet?.reloadData()                           TODO
-        })
+        extendSelectionAction { _, _ in
+            actionSheet.reloadData()
+        }
     }
     
     open func extendSelectionAction(with action: @escaping CellAction) {