+++ Sheeeeeeeeet/Items/Items/ActionSheetSingleSelectItem.swift
@@ -47,8 +47,8 @@ open class ActionSheetSingleSelectItem: ActionSheetSelectItem {
         self.appearance = ActionSheetSingleSelectItemAppearance(copy: appearance.singleSelectItem)
     }
     
-    open override func handleTap(in actionSheet: ActionSheet, cell: UITableViewCell? = nil) {
-        super.handleTap(in: actionSheet, cell: cell)
+    open override func handleTap(in actionSheet: ActionSheet) {
+        super.handleTap(in: actionSheet)
         let items = actionSheet.items.compactMap { $0 as? ActionSheetSingleSelectItem }
         let deselectItems = items.filter { $0.group == group }
         deselectItems.forEach { $0.isSelected = false }